requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
autobahn==19.11.2
Automat==20.2.0
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
Django==4.1.3
django-appconf==1.0.5
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.13.1
djlint==1.3.0
drf-spectacular==0.22.1
drf-writable-nested==0.6.3
enum34==1.1.10
et-xmlfile==1.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
inflection==0.5.1
jsonschema==4.6.1
kiteconnect==4.1.0
kombu==5.2.4
mypy-extensions==0.4.3
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
pathspec==0.9.0
pilkit==2.0
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
pyOpenSSL==22.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==305
PyYAML==6.0
regex==2022.6.2
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.0
zope.interface==5.4.0
